requirements.txt
google-genai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
trendspyg>=0.3.0
feedparser>=6.0.0
APScheduler>=3.10.0
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
websocket-client>=1.6.0
fastapi>=0.115.0
uvicorn>=0.34.0
pyngrok>=7.0.0
pyyaml>=6.0
sqlalchemy>=2.0
aiosqlite>=0.20
aiomysql>=0.2.0
cryptography>=42.0
alembic>=1.13
bcrypt>=5.0.0
httpx>=0.27.0
google-cloud-storage>=2.14.0
